feat(backend): 이벤트 목록/상세 조회 API 개발 - 수정1
- EventPace 생성자 인자 Byte -> Integer변경(tinyint->int)
- api 매핑 주소 /api 제거로 media와 통일

+++ backend/main/src/main/java/com/kt/onrace/domain/event/entity/EventPace.java
@@ -33,7 +33,7 @@ public class EventPace extends BaseEntity {
 	private int minutes = 0;
 
 	@Builder
-	public EventPace(EventCourse eventCourse, String name, Byte hour, Byte minutes) {
+	public EventPace(EventCourse eventCourse, String name, Integer hour, Integer minutes) {
 		this.eventCourse = eventCourse;
 		this.name = name;
 		this.hour = hour != null ? hour : 0;
